Ingestion/requirements.txt
pandas 
s3fs
SQLalchemy 
psycopg2-binary
datetime
python-dotenv 